LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_sequence_at_implicit_release_on_click-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Event: Event sequence at implicit release on click

Follow the test instructions with mouse. If you don't have the device skip it.

When a captured pointer is implicitly released after a click, the boundary events should follow the lostpointercapture event.

Click or tap on Black.
Click or tap on Green.

Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN mouse Event sequence at implicit release on click
